Fixed issue with undefined killer variable crashing bds

+++ Packs/Skyblock/scripts/skyblock.js
@@ -62,7 +62,7 @@ world.afterEvents.entityDie.subscribe((event) => {//handles on death
 	else{
 		if(typeof event.damageSource !== "undefined"){
 			let killer = event.damageSource.damagingEntity//gets the player who killed the entity
-			if (killer.typeId  == "minecraft:player"){//if the damage source was a player
+			if (killer && killer.typeId  == "minecraft:player"){//if the damage source was a player
 				world.scoreboard.getObjective("Kill Counter").addScore(killer,1)//add to kill counter
 			}
 		}
